ConfigAbstractFactory::canCreate currently returns true on service names, which it can not create. So has(name) followed by get(name) may throw.
Abstract factories (must) all assume, that configuration does not change
between canCreate and __invoke. Othewise canCreate would be meaningless.
(Reminder, deprecate canCreate asap and extend AbstractFactory::__invoke
that way that it returns null if not applicable (this again will break
an assumption they made, that null could be validly produced by a
factory)). 

This commit is the first step to cache ConfigAbstractFactory
configuration between canCreate and __invoke. Nothing to be done fast
track, because this thing is recursive as hell. This first step passes
all the tests, but is still not as fast as possible.

Additionally to just caching configuration this commit provides more
detailled error exceptions. Tests had be to be justified to accomodate
that correctly.